Bugs fixed

+++ Canteen/views.py
@@ -7,11 +7,12 @@ from datetime import datetime
 from Login.models import User_class
 from django.conf import settings
 from django.core.mail import send_mail
+import pytz
 
 def Student_Place_Order(request):
     if request.user.is_authenticated:
         if request.user.designation == "Student":
-            menu = Menu.objects.all()
+            menu = Menu.objects.all().exclude(Item_Name = "")
             if request.method == "POST":
                 quantity = int(request.POST.get("quantity"))
                 idt = request.POST.get("submit")
@@ -51,8 +52,9 @@ def Student_Cart(request):
             if request.method == "POST":
                 if "order_validation1" in request.POST:
                     order = request.POST.get("order_validation1")
-                    req = Order.objects.filter(id=int(order))[0]
-                    req.delete()
+                    if Order.objects.filter(id=int(order)):
+                        req = Order.objects.filter(id=int(order))[0]
+                        req.delete()
                     messages.success(
                         request, "Your Cart Item has been successfully removed"
                     )
@@ -151,8 +153,10 @@ def Owner_New_Order(request):
                     
                     username = order.User_Name
                     name = User_class.objects.filter(username=username)[0].name
+                    dt = order.Order_Date_Time.astimezone(pytz.timezone('Asia/Kolkata'))
+                    dt = dt.strftime("%Y-%m-%d %H:%M:%S")
                     subject = "Order Rejected"
-                    message = f"Dear {name}, Your order, made on {order.Order_Date_Time}, with items {order.Item_Name} has been rejected. We deeply regret this. Contact the canteen manager."
+                    message = f"Dear {name}, Your order, made on {dt}, with items {order.Item_Name} has been rejected. We deeply regret this. Contact the canteen manager."
                     email_from = settings.EMAIL_HOST_USER
                     recipient_list = [
                         f"{username}@iitk.ac.in",
@@ -284,8 +288,9 @@ def Owner_Modify_Menu(request):
                             extra_items.Item_Name = item
                             extra_items.Price = price
                             extra_items.save()
-                    extra_items_del = Menu.objects.filter(id=idt)[0]
-                    extra_items_del.delete()
+                    if Menu.objects.filter(id=idt):
+                        extra_items_del = Menu.objects.filter(id=idt)[0]
+                        extra_items_del.delete()
                     return render(
                         request,
                         "Owner_Modify_Menu.html",
